@@ -14,7 +14,7 @@ authors = [
 description = "Software Package for Analyzing Relativistic Kinematics in Collision eXperiments"
 readme = "README.md"
 license = {file = "LICENSE"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Intended Audience :: Developers",
