@@ -10,7 +10,8 @@
 from sparkx.Filter import *
 import numpy as np
 from abc import ABC, abstractmethod
-from typing import List, Union, Tuple, Optional
+from sparkx.Particle import Particle
+from typing import List, Union, Tuple, Optional, Any
 from sparkx.loader.BaseLoader import BaseLoader
 
 
@@ -61,6 +62,8 @@ class BaseStorer(ABC):
         Keep strange particles only
     pT_cut:
         Apply pT cut to all particles
+    mT_cut:
+        Apply mT cut to all particles
     rapidity_cut:
         Apply rapidity cut to all particles
     pseudorapidity_cut:
@@ -74,7 +77,7 @@ class BaseStorer(ABC):
     """
 
     def __init__(
-        self, path: Union[str, List[List["Particle"]]], **kwargs
+        self, path: Union[str, List[List[Particle]]], **kwargs: Any
     ) -> None:
         """
         Parameters
@@ -94,16 +97,100 @@ class BaseStorer(ABC):
         self.loader_: Optional[BaseLoader] = None
         self.num_output_per_event_: Optional[np.ndarray] = None
         self.num_events_: Optional[int] = None
-        self.particle_list_: Optional[List] = None
+        self.particle_list_: List[List[Particle]] = [[]]
+        self.custom_attr_list: List = []
         self.create_loader(path)
         if self.loader_ is not None:
             (
                 self.particle_list_,
                 self.num_events_,
                 self.num_output_per_event_,
+                self.custom_attr_list
             ) = self.loader_.load(**kwargs)
         else:
             raise ValueError("Loader has not been created properly")
+        
+    def __add__(self, other: "BaseStorer") -> "BaseStorer":
+        """
+        Adds two BaseStorer objects by combining their particle lists and updating num_output_per_event accordingly.
+
+        This method ensures that both objects are instances of the same class before combining them. If the objects
+        are not of the same class, a TypeError is raised.
+
+        Parameters
+        ----------
+        other : BaseStorer
+            The other BaseStorer object to be added.
+
+        Raises
+        ------
+        TypeError
+            If the other object is not an instance of BaseStorer or if the objects are not of the same class.
+
+        Returns
+        -------
+        BaseStorer
+            A new BaseStorer object with combined particle lists and updated num_output_per_event.
+        """
+        if not isinstance(other, BaseStorer):
+            raise TypeError("Can only add BaseStorer objects")
+        
+        # Ensure that both instances are of the same class
+        if type(self) is not type(other):
+            raise TypeError("Can only add objects of the same class")
+
+        combined_particle_list: list = self.particle_list_ + other.particle_list_
+
+        # Ensure num_output_per_event_ is not None
+        if self.num_output_per_event_ is None:
+            self.num_output_per_event_ = np.empty((0, 2), dtype=int)
+        if other.num_output_per_event_ is None:
+            other.num_output_per_event_ = np.empty((0, 2), dtype=int)
+        if self.num_events_ is None:
+            self.num_events_ = 0
+        if other.num_events_ is None:
+            other.num_events_ = 0
+
+        combined_num_output_per_event: np.ndarray = np.concatenate(
+            (self.num_output_per_event_, other.num_output_per_event_)
+        )
+
+        # Adjust event_number for the parts that originally belonged to other
+        combined_num_output_per_event[self.num_events_:, 0] += self.num_events_
+
+        combined_storer: BaseStorer = self.__class__.__new__(self.__class__)
+        combined_storer.__dict__.update(self.__dict__)  # Inherit all properties from self
+        combined_storer._update_after_merge(other)
+        combined_storer.particle_list_ = combined_particle_list
+        combined_storer.num_output_per_event_ = combined_num_output_per_event
+        combined_storer.num_events_ = self.num_events_ + other.num_events_
+        combined_storer.loader_ = None  # Loader is not applicable for combined object
+
+        return combined_storer
+    
+    @abstractmethod
+    def _update_after_merge(self, other: "BaseStorer") -> None:
+        """
+        Updates the attributes of the current instance after merging with another BaseStorer object.
+
+        This method should be implemented by subclasses to update the attributes of the current instance after merging
+        with another BaseStorer object. The method raises a NotImplementedError if it is not overridden by a subclass.
+
+        Parameters
+        ----------
+        other : BaseStorer
+            The other BaseStorer object that was merged with the current instance.
+
+        Raises
+        ------
+        NotImplementedError
+            If the method is not implemented by a subclass.
+
+        Returns
+        -------
+        None
+        """
+        raise NotImplementedError("This method is not implemented yet")
 
     @abstractmethod
     def create_loader(self, arg: Union[str, List[List["Particle"]]]) -> None:
@@ -114,10 +201,10 @@ class BaseStorer(ABC):
         Returns a numpy array containing the event number (starting with 1)
         and the corresponding number of particles created in this event as
 
-        num_output_per_event[event_n, number_of_particles_in_event_n]
+        :code:`num_output_per_event[event_n, number_of_particles_in_event_n]`
 
-        num_output_per_event is updated with every manipulation e.g. after
-        applying cuts.
+        :code:`num_output_per_event` is updated with every manipulation e.g. 
+        after applying cuts.
 
         Returns
         -------
@@ -129,23 +216,22 @@ class BaseStorer(ABC):
 
     def num_events(self) -> Optional[int]:
         """
-        Returns the number of events in particle_list
+        Returns the number of events in :code:`particle_list`.
 
-        num_events is updated with every manipulation e.g. after
+        :code:`num_events` is updated with every manipulation e.g. after
         applying cuts.
 
         Returns
         -------
         num_events_ : int
-            Number of events in particle_list
+            Number of events in :code:`particle_list`
         """
         return self.num_events_
 
     def particle_objects_list(self) -> Optional[List]:
         """
         Returns a nested python list containing all particles from
-        the data as particle objects
-        from Particle:
+        the data as particle objects from :code:`Particle`:
 
            | Single Event:    [particle_object]
            | Multiple Events: [event][particle_object]
@@ -153,7 +239,7 @@ class BaseStorer(ABC):
         Returns
         -------
         particle_list_ : list
-            List of particle objects from Particle
+            List of particle objects from :code:`Particle`
         """
         return self.particle_list_
 
@@ -173,7 +259,6 @@ class BaseStorer(ABC):
         -------
         list
             Nested list containing the current data
-
         """
         num_events = self.num_events_
 
@@ -207,14 +292,13 @@ class BaseStorer(ABC):
 
     def charged_particles(self) -> "BaseStorer":
         """
-        Keep only charged particles in particle_list
+        Keep only charged particles in :code:`particle_list`.
 
         Returns
         -------
         self : BaseStorer object
             Containing charged particles in every event only
         """
-
         self.particle_list_ = charged_particles(self.particle_list_)
 
         self._update_num_output_per_event_after_filter()
@@ -222,30 +306,14 @@ class BaseStorer(ABC):
 
     def uncharged_particles(self) -> "BaseStorer":
         """
-        Keep only uncharged particles in particle_list
+        Keep only uncharged particles in :code:`particle_list`.
 
         Returns
         -------
         self : BaseStorer object
             Containing uncharged particles in every event only
         """
-
         self.particle_list_ = uncharged_particles(self.particle_list_)
-        self._update_num_output_per_event_after_filter()
-
-        return self
-
-    def strange_particles(self) -> "BaseStorer":
-        """
-        Keep only strange particles in particle_list
-
-        Returns
-        -------
-        self : BaseStorer object
-            Containing strange particles in every event only
-        """
-
-        self.particle_list_ = strange_particles(self.particle_list_)
         self._update_num_output_per_event_after_filter()
 
         return self
@@ -254,7 +322,7 @@ class BaseStorer(ABC):
         self, pdg_list: Union[int, Union[Tuple[int], List[int], np.ndarray]]
     ) -> "BaseStorer":
         """
-        Keep only particle species given by their PDG ID in every event
+        Keep only particle species given by their PDG ID in every event.
 
         Parameters
         ----------
@@ -268,10 +336,9 @@ class BaseStorer(ABC):
         Returns
         -------
         self : BaseStorer object
-            Containing only particle species specified by pdg_list for every event
-
+            Containing only particle species specified by :code:`pdg_list` for 
+            every event
         """
-
         self.particle_list_ = particle_species(self.particle_list_, pdg_list)
         self._update_num_output_per_event_after_filter()
 
@@ -281,8 +348,8 @@ class BaseStorer(ABC):
         self, pdg_list: Union[int, Union[Tuple[int], List[int], np.ndarray]]
     ) -> "BaseStorer":
         """
-        Remove particle species from particle_list by their PDG ID in every
-        event
+        Remove particle species from :code:`particle_list` by their PDG ID in 
+        every event.
 
         Parameters
         ----------
@@ -297,9 +364,7 @@ class BaseStorer(ABC):
         -------
         self : BaseStorer object
             Containing all but the specified particle species in every event
-
         """
-
         self.particle_list_ = remove_particle_species(
             self.particle_list_, pdg_list
         )
@@ -309,14 +374,13 @@ class BaseStorer(ABC):
 
     def participants(self) -> "BaseStorer":
         """
-        Keep only participants in particle_list
+        Keep only participants in :code:`particle_list`.
 
         Returns
         -------
         self : BaseStorer object
             Containing participants in every event only
         """
-
         self.particle_list_ = participants(self.particle_list_)
         self._update_num_output_per_event_after_filter()
 
@@ -324,14 +388,13 @@ class BaseStorer(ABC):
 
     def spectators(self) -> "BaseStorer":
         """
-        Keep only spectators in particle_list
+        Keep only spectators in :code:`particle_list`.
 
         Returns
         -------
         self : BaseStorer object
             Containing spectators in every event only
         """
-
         self.particle_list_ = spectators(self.particle_list_)
         self._update_num_output_per_event_after_filter()
 
@@ -346,7 +409,8 @@ class BaseStorer(ABC):
         Parameters
         ----------
         minimum_event_energy : int or float
-            The minimum event energy threshold. Should be a positive integer or float.
+            The minimum event energy threshold. Should be a positive integer or 
+            float.
 
         Returns
         -------
@@ -357,11 +421,12 @@ class BaseStorer(ABC):
         Raises
         ------
         TypeError
-            If the minimum_event_energy parameter is not an integer or float.
+            If the :code:`minimum_event_energy` parameter is not an integer or 
+            float.
         ValueError
-            If the minimum_event_energy parameter is less than or equal to 0.
+            If the :code:`minimum_event_energy` parameter is less than or 
+            equal to 0.
         """
-
         self.particle_list_ = lower_event_energy_cut(
             self.particle_list_, minimum_event_energy
         )
@@ -397,20 +462,48 @@ class BaseStorer(ABC):
 
         return self
 
+    def mT_cut(
+        self, cut_value_tuple: Tuple[Union[float, None], Union[float, None]]
+    ) -> "BaseStorer":
+        """
+        Apply transverse mass cut to all events by passing an acceptance
+        range by :code:`cut_value_tuple`. All particles outside this range will
+        be removed.
+
+        Parameters
+        ----------
+        cut_value_tuple : tuple
+            Tuple with the upper and lower limits of the mT acceptance
+            range :code:`(cut_min, cut_max)`. If one of the limits is not
+            required, set it to :code:`None`, i.e. :code:`(None, cut_max)`
+            or :code:`(cut_min, None)`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing only particles complying with the transverse mass
+            cut for all events
+        """
+
+        self.particle_list_ = mT_cut(self.particle_list_, cut_value_tuple)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
     def rapidity_cut(
         self, cut_value: Union[float, Tuple[float, float]]
     ) -> "BaseStorer":
         """
         Apply rapidity cut to all events and remove all particles with rapidity
-        not complying with cut_value
+        not complying with :code:`cut_value`.
 
         Parameters
         ----------
         cut_value : float
             If a single value is passed, the cut is applied symmetrically
             around 0.
-            For example, if :code:`cut_value = 1`, only particles with rapidity in
-            :code:`[-1.0, 1.0]` are kept.
+            For example, if :code:`cut_value = 1`, only particles with rapidity 
+            in :code:`[-1.0, 1.0]` are kept.
 
         cut_value : tuple
             To specify an asymmetric acceptance range for the rapidity
@@ -422,7 +515,6 @@ class BaseStorer(ABC):
             Containing only particles complying with the rapidity cut
             for all events
         """
-
         self.particle_list_ = rapidity_cut(self.particle_list_, cut_value)
         self._update_num_output_per_event_after_filter()
 
@@ -433,7 +525,7 @@ class BaseStorer(ABC):
     ) -> "BaseStorer":
         """
         Apply pseudo-rapidity cut to all events and remove all particles with
-        pseudo-rapidity not complying with cut_value
+        pseudo-rapidity not complying with :code:`cut_value`.
 
         Parameters
         ----------
@@ -459,12 +551,13 @@ class BaseStorer(ABC):
 
         return self
 
-    def spatial_rapidity_cut(
+    def spacetime_rapidity_cut(
         self, cut_value: Union[float, Tuple[float, float]]
     ) -> "BaseStorer":
         """
         Apply spacetime rapidity (space-time rapidity) cut to all events and
-        remove all particles with spcetime rapidity not complying with cut_value
+        remove all particles with spacetime rapidity not complying with 
+        cut_value.
 
         Parameters
         ----------
@@ -492,16 +585,19 @@ class BaseStorer(ABC):
 
         return self
 
-    def multiplicity_cut(self, min_multiplicity: int) -> "BaseStorer":
+    def multiplicity_cut(
+        self, cut_value_tuple: Tuple[Union[float, None], Union[float, None]]
+        ) -> "BaseStorer":
         """
-        Apply multiplicity cut. Remove all events with a multiplicity lower
-        than min_multiplicity
+        Apply multiplicity cut. Remove all events with a multiplicity not
+        complying with cut_value.
 
         Parameters
         ----------
-        min_multiplicity : int
-            Lower bound for multiplicity. If the multiplicity of an event is
-            lower than min_multiplicity, this event is discarded.
+        cut_value_tuple : tuple
+            Upper and lower bound for multiplicity. If the multiplicity of an event is
+            not in this range, the event is discarded. The range is inclusive on the 
+            lower bound and exclusive on the upper bound.
 
         Returns
         -------
@@ -510,13 +606,15 @@ class BaseStorer(ABC):
         """
 
         self.particle_list_ = multiplicity_cut(
-            self.particle_list_, min_multiplicity
+            self.particle_list_, cut_value_tuple
         )
         self._update_num_output_per_event_after_filter()
 
         return self
 
-    def spacetime_cut(self, dim, cut_value_tuple: Tuple[float, float]):
+    def spacetime_cut(
+        self, dim: str, cut_value_tuple: Tuple[float, float]
+    ) -> "BaseStorer":
         """
         Apply spacetime cut to all events by passing an acceptance range by
         :code:`cut_value_tuple`. All particles outside this range will
@@ -536,11 +634,206 @@ class BaseStorer(ABC):
         Returns
         -------
         self : BaseStorer object
-            Containing only particles complying with the spacetime cut for all events
+            Containing only particles complying with the spacetime cut for all 
+            events
         """
         self.particle_list_ = spacetime_cut(
             self.particle_list_, dim, cut_value_tuple
         )
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def particle_status(
+        self, status_list: Union[int, Tuple[int, ...], List[int], np.ndarray]
+    ) -> "BaseStorer":
+        """
+        Keep only particles with a given particle status.
+
+        Parameters
+        ----------
+        status_list : int
+            To keep a particles with a single status only, pass a single status
+
+        status_list : tuple/list/array
+            To keep hadrons with different hadron status, pass a tuple or list
+            or array
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing only hadrons with status specified by 
+            :code:`status_list` for every event
+        """
+        self.particle_list_ = particle_status(self.particle_list_, status_list)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_hadrons(self) -> "BaseStorer":
+        """
+        Keep only hadrons in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons in every event only
+        """
+        self.particle_list_ = keep_hadrons(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_leptons(self) -> "BaseStorer":
+        """
+        Keep only leptons in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing leptons in every event only
+        """
+        self.particle_list_ = keep_leptons(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_quarks(self) -> "BaseStorer":
+        """
+        Keep only quarks in the :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing quarks in every event only
+        """
+        self.particle_list_ = keep_quarks(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_mesons(self) -> "BaseStorer":
+        """
+        Keep only mesons in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing mesons in every event only
+        """
+        self.particle_list_ = keep_mesons(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_baryons(self) -> "BaseStorer":
+        """
+        Keep only baryons in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing baryons in every event only
+        """
+        self.particle_list_ = keep_baryons(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_up(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing up quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing up quarks in every event only
+        """
+        self.particle_list_ = keep_up(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_down(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing down quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing down quarks in every event only
+        """
+        self.particle_list_ = keep_down(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_strange(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing strange quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing strange quarks in every event only
+        """
+        self.particle_list_ = keep_strange(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_charm(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing charm quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing charm quarks in every event only
+        """
+        self.particle_list_ = keep_charm(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_bottom(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing bottom quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing bottom quarks in every event only
+        """
+        self.particle_list_ = keep_bottom(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def keep_top(self) -> "BaseStorer":
+        """
+        Keep only hadrons containing top quarks in :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing hadrons containing top quarks in every event only
+        """
+        self.particle_list_ = keep_top(self.particle_list_)
+        self._update_num_output_per_event_after_filter()
+
+        return self
+
+    def remove_photons(self) -> "BaseStorer":
+        """
+        Remove photons from :code:`particle_list`.
+
+        Returns
+        -------
+        self : BaseStorer object
+            Containing all but photons in every event
+        """
+        self.particle_list_ = remove_photons(self.particle_list_)
         self._update_num_output_per_event_after_filter()
 
         return self
@@ -550,16 +843,19 @@ class BaseStorer(ABC):
             raise ValueError("num_output_per_event_ is not set")
         if self.particle_list_ is None:
             raise ValueError("particle_list_ is not set")
-
         if self.num_output_per_event_.ndim == 1:
             # Handle the case where num_output_per_event_ is a one-dimensional array
             self.num_output_per_event_[1] = len(self.particle_list_[0])
         elif self.num_output_per_event_.ndim == 2:
             # Handle the case where num_output_per_event_ is a two-dimensional array
+            updated_num_output_per_event : np.ndarray = np.ndarray((len(self.particle_list_),2), dtype=int)
             for event in range(len(self.particle_list_)):
-                self.num_output_per_event_[event][1] = len(
+                updated_num_output_per_event[event][0] = event + self.num_output_per_event_[0][0]
+                updated_num_output_per_event[event][1] = len(
                     self.particle_list_[event]
                 )
+            self.num_output_per_event_ = updated_num_output_per_event
+            self.num_events_ = len(self.particle_list_)
         else:
             raise ValueError(
                 "num_output_per_event_ has an unexpected number of dimensions"
@@ -571,8 +867,8 @@ class BaseStorer(ABC):
         Prints the particle lists to a specified file.
 
         This method should be implemented by subclasses to print the particle
-        lists to the specified output file. The method raises a NotImplementedError
-        if it is not overridden by a subclass.
+        lists to the specified output file. The method raises a 
+        :code:`NotImplementedError` if it is not overridden by a subclass.
 
         Parameters
         ----------
