@@ -10,9 +10,15 @@
 import numpy as np
 from sparkx.Particle import Particle
 import warnings
+from typing import Tuple, List, Union, Optional
 
 
-def __ensure_tuple_is_valid_else_raise_error(value_tuple, allow_none=False):
+def __ensure_tuple_is_valid_else_raise_error(
+    value_tuple: Tuple[
+        Optional[Union[int, float]], Optional[Union[float, int]]
+    ],
+    allow_none: bool = False,
+) -> None:
     """
     Validates a tuple for specific conditions.
 
@@ -63,9 +69,11 @@ def __ensure_tuple_is_valid_else_raise_error(value_tuple, allow_none=False):
             raise ValueError("At least one cut limit must be set to a number")
 
 
-def charged_particles(particle_list):
+def charged_particles(
+    particle_list: List[List[Particle]],
+) -> List[List[Particle]]:
     """
-    Keep only charged particles in particle_list.
+    Keep only charged particles in :code:`particle_list`.
 
     Parameters
     ----------
@@ -86,9 +94,11 @@ def charged_particles(particle_list):
     return particle_list
 
 
-def uncharged_particles(particle_list):
+def uncharged_particles(
+    particle_list: List[List[Particle]],
+) -> List[List[Particle]]:
     """
-    Keep only uncharged particles in particle_list.
+    Keep only uncharged particles in :code:`particle_list`.
 
     Parameters
     ----------
@@ -109,30 +119,10 @@ def uncharged_particles(particle_list):
     return particle_list
 
 
-def strange_particles(particle_list):
-    """
-    Keep only strange particles in particle_list.
-
-    Parameters
-    ----------
-    particle_list:
-        List with lists containing particle objects for the events
-
-    Returns
-    -------
-    list of lists
-        Filtered list of lists containing particle objects for each event
-    """
-    for i in range(0, len(particle_list)):
-        particle_list[i] = [
-            elem
-            for elem in particle_list[i]
-            if elem.is_strange() and not np.isnan(elem.is_strange())
-        ]
-    return particle_list
-
-
-def particle_species(particle_list, pdg_list):
+def particle_species(
+    particle_list: List[List[Particle]],
+    pdg_list: Union[int, Tuple[int, ...], List[int], np.ndarray],
+) -> List[List[Particle]]:
     """
     Keep only particle species given by their PDG ID in every event.
 
@@ -172,7 +162,6 @@ def particle_species(particle_list, pdg_list):
 
     if isinstance(pdg_list, (int, float, str, np.integer)):
         pdg_list = int(pdg_list)
-
         for i in range(0, len(particle_list)):
             particle_list[i] = [
                 elem
@@ -199,9 +188,12 @@ def particle_species(particle_list, pdg_list):
     return particle_list
 
 
-def remove_particle_species(particle_list, pdg_list):
+def remove_particle_species(
+    particle_list: List[List[Particle]],
+    pdg_list: Union[int, Tuple[int, ...], List[int], np.ndarray],
+) -> List[List[Particle]]:
     """
-    Remove particle species from particle_list by their PDG ID in every
+    Remove particle species from :code:`particle_list` by their PDG ID in every
     event.
 
     Parameters
@@ -240,7 +232,6 @@ def remove_particle_species(particle_list, pdg_list):
 
     if isinstance(pdg_list, (int, float, str, np.integer)):
         pdg_list = int(pdg_list)
-
         for i in range(0, len(particle_list)):
             particle_list[i] = [
                 elem
@@ -267,9 +258,9 @@ def remove_particle_species(particle_list, pdg_list):
     return particle_list
 
 
-def participants(particle_list):
+def participants(particle_list: List[List[Particle]]) -> List[List[Particle]]:
     """
-    Keep only participants in particle_list.
+    Keep only participants in :code:`particle_list`.
 
     Parameters
     ----------
@@ -291,9 +282,9 @@ def participants(particle_list):
     return particle_list
 
 
-def spectators(particle_list):
+def spectators(particle_list: List[List[Particle]]) -> List[List[Particle]]:
     """
-    Keep only spectators in particle_list.
+    Keep only spectators in :code:`particle_list`.
 
     Parameters
     ----------
@@ -315,7 +306,9 @@ def spectators(particle_list):
     return particle_list
 
 
-def lower_event_energy_cut(particle_list, minimum_event_energy):
+def lower_event_energy_cut(
+    particle_list: List[List[Particle]], minimum_event_energy: Union[int, float]
+) -> List[List[Particle]]:
     """
     Filters out events with total energy lower than a threshold.
     Events with smaller energies are removed and not kept as empty events.
@@ -326,7 +319,8 @@ def lower_event_energy_cut(particle_list, minimum_event_energy):
         List with lists containing particle objects for the events
 
     minimum_event_energy : int or float
-        The minimum event energy threshold. Should be a positive integer or float.
+        The minimum event energy threshold. Should be a positive integer or 
+        float.
 
     Returns
     -------
@@ -336,11 +330,11 @@ def lower_event_energy_cut(particle_list, minimum_event_energy):
     Raises
     ------
     TypeError
-        If the minimum_event_energy parameter is not an integer or float.
+        If the :code:`minimum_event_energy` parameter is not an integer or float.
     ValueError
-        If the minimum_event_energy parameter is less than or equal to 0.
+        If the :code:`minimum_event_energy` parameter is less than or equal to 0.
     ValueError
-        If the minimum_event_energy parameter is NaN.
+        If the :code:`minimum_event_energy` parameter is :code:`NaN`.
     """
     if not isinstance(minimum_event_energy, (int, float)):
         raise TypeError(
@@ -369,7 +363,11 @@ def lower_event_energy_cut(particle_list, minimum_event_energy):
     return particle_list
 
 
-def spacetime_cut(particle_list, dim, cut_value_tuple):
+def spacetime_cut(
+    particle_list: List[List[Particle]],
+    dim: str,
+    cut_value_tuple: Tuple[Optional[float], Optional[float]],
+) -> List[List[Particle]]:
     """
     Apply spacetime cut to all events by passing an acceptance range by
     ::code`cut_value_tuple`. All particles outside this range will
@@ -451,9 +449,12 @@ def spacetime_cut(particle_list, dim, cut_value_tuple):
     return updated_particle_list
 
 
-def pT_cut(particle_list, cut_value_tuple):
+def pT_cut(
+    particle_list: List[List[Particle]],
+    cut_value_tuple: Tuple[Optional[float], Optional[float]],
+) -> List[List[Particle]]:
     """
-    Apply p_t cut to all events by passing an acceptance range by
+    Apply pT cut to all events by passing an acceptance range by
     ::code`cut_value_tuple`. All particles outside this range will
     be removed.
 
@@ -521,7 +522,10 @@ def pT_cut(particle_list, cut_value_tuple):
     return updated_particle_list
 
 
-def mT_cut(particle_list, cut_value_tuple):
+def mT_cut(
+    particle_list: List[List[Particle]],
+    cut_value_tuple: Tuple[Optional[float], Optional[float]],
+) -> List[List[Particle]]:
     """
     Apply transverse mass cut to all events by passing an acceptance range by
     ::code`cut_value_tuple`. All particles outside this range will
@@ -588,7 +592,10 @@ def mT_cut(particle_list, cut_value_tuple):
     return updated_particle_list
 
 
-def rapidity_cut(particle_list, cut_value):
+def rapidity_cut(
+    particle_list: List[List[Particle]],
+    cut_value: Union[int, float, Tuple[float, float]],
+) -> List[List[Particle]]:
     """
     Apply rapidity cut to all events and remove all particles with rapidity
     not complying with cut_value.
@@ -601,12 +608,12 @@ def rapidity_cut(particle_list, cut_value):
     cut_value : float
         If a single value is passed, the cut is applied symmetrically
         around 0.
-        For example, if cut_value = 1, only particles with rapidity in
+        For example, if :code:`cut_value = 1`, only particles with rapidity in
         [-1.0, 1.0] are kept.
 
     cut_value : tuple
         To specify an asymmetric acceptance range for the rapidity
-        of particles, pass a tuple (cut_min, cut_max)
+        of particles, pass a tuple :code:`(cut_min, cut_max)`
 
     Returns
     -------
@@ -657,10 +664,13 @@ def rapidity_cut(particle_list, cut_value):
     return updated_particle_list
 
 
-def pseudorapidity_cut(particle_list, cut_value):
+def pseudorapidity_cut(
+    particle_list: List[List[Particle]],
+    cut_value: Union[int, float, Tuple[float, float]],
+) -> List[List[Particle]]:
     """
     Apply pseudo-rapidity cut to all events and remove all particles with
-    pseudo-rapidity not complying with cut_value.
+    pseudo-rapidity not complying with :code:`cut_value`.
 
     Parameters
     ----------
@@ -670,12 +680,12 @@ def pseudorapidity_cut(particle_list, cut_value):
     cut_value : float
         If a single value is passed, the cut is applied symmetrically
         around 0.
-        For example, if cut_value = 1, only particles with pseudo-rapidity
-        in [-1.0, 1.0] are kept.
+        For example, if :code:`cut_value = 1`, only particles with 
+        pseudo-rapidity in [-1.0, 1.0] are kept.
 
     cut_value : tuple
         To specify an asymmetric acceptance range for the pseudo-rapidity
-        of particles, pass a tuple (cut_min, cut_max)
+        of particles, pass a tuple :code:`(cut_min, cut_max)`
 
     Returns
     -------
@@ -726,10 +736,13 @@ def pseudorapidity_cut(particle_list, cut_value):
     return updated_particle_list
 
 
-def spacetime_rapidity_cut(particle_list, cut_value):
+def spacetime_rapidity_cut(
+    particle_list: List[List[Particle]],
+    cut_value: Union[int, float, Tuple[float, float]],
+) -> List[List[Particle]]:
     """
     Apply space-time rapidity cut to all events and remove all particles with
-    space-time rapidity not complying with cut_value.
+    space-time rapidity not complying with :code:`cut_value`.
 
     Parameters
     ----------
@@ -739,12 +752,12 @@ def spacetime_rapidity_cut(particle_list, cut_value):
     cut_value : float
         If a single value is passed, the cut is applied symmetrically
         around 0.
-        For example, if cut_value = 1, only particles with spacetime rapidity
-        in [-1.0, 1.0] are kept.
+        For example, if :code:`cut_value = 1`, only particles with spacetime 
+        rapidity in [-1.0, 1.0] are kept.
 
     cut_value : tuple
         To specify an asymmetric acceptance range for the space-time rapidity
-        of particles, pass a tuple (cut_min, cut_max)
+        of particles, pass a tuple :code:`(cut_min, cut_max)`
 
     Returns
     -------
@@ -795,34 +808,60 @@ def spacetime_rapidity_cut(particle_list, cut_value):
     return updated_particle_list
 
 
-def multiplicity_cut(particle_list, min_multiplicity):
+def multiplicity_cut(
+    particle_list: List[List[Particle]], cut_value_tuple: tuple
+) -> List[List[Particle]]:
     """
-    Apply multiplicity cut. Remove all events with a multiplicity lower
-    than min_multiplicity.
+    Apply multiplicity cut. Remove all events with a multiplicity not complying
+    with cut_value_tuple.
 
     Parameters
     ----------
     particle_list:
         List with lists containing particle objects for the events
 
-    min_multiplicity : int
-        Lower bound for multiplicity. If the multiplicity of an event is
-        lower than min_multiplicity, this event is discarded.
+   cut_value_tuple : tuple
+        Upper and lower bound for multiplicity. If the multiplicity of an event is
+        not in this range, the event is discarded. The range is inclusive on the 
+        lower bound and exclusive on the upper bound.
 
     Returns
     -------
     list of lists
         Filtered list of lists containing particle objects for each event
     """
-    if not isinstance(min_multiplicity, int):
-        raise TypeError("Input value for multiplicity cut must be an int")
-    if min_multiplicity < 0:
-        raise ValueError("Minimum multiplicity must >= 0")
+    if not isinstance(cut_value_tuple, tuple):
+        raise TypeError(
+            "Input value must be a tuple containing either "
+            + "positive numbers or None of length two"
+        )
+
+    __ensure_tuple_is_valid_else_raise_error(cut_value_tuple, allow_none=True)
+
+    # Check if the cut limits are positive if they are not None
+    if (cut_value_tuple[0] is not None and cut_value_tuple[0] < 0) or (
+        cut_value_tuple[1] is not None and cut_value_tuple[1] < 0
+    ):
+        raise ValueError("The cut limits must be positive or None")
+
+    if cut_value_tuple[0] is None:
+        lower_cut = float("-inf")
+    else:
+        lower_cut = cut_value_tuple[0]
+
+    if cut_value_tuple[1] is None:
+        upper_cut = float("inf")
+    else:
+        upper_cut = cut_value_tuple[1]
+
+    # Ensure cut values are in the correct order
+    lim_max = max(upper_cut, lower_cut)
+    lim_min = min(upper_cut, lower_cut)
 
     idx_keep_event = []
     for idx, event_particles in enumerate(particle_list):
         multiplicity = len(event_particles)
-        if multiplicity >= min_multiplicity:
+        if multiplicity >= lower_cut and multiplicity < upper_cut:
             idx_keep_event.append(idx)
 
     particle_list = [particle_list[idx] for idx in idx_keep_event]
@@ -830,7 +869,10 @@ def multiplicity_cut(particle_list, min_multiplicity):
     return particle_list
 
 
-def particle_status(particle_list, status_list):
+def particle_status(
+    particle_list: List[List[Particle]],
+    status_list: Union[int, np.ndarray, List[int], Tuple[int, ...]],
+) -> List[List[Particle]]:
     """
     Keep only particles with a given particle status.
 
@@ -884,4 +926,287 @@ def particle_status(particle_list, status_list):
 
     particle_list = updated_particle_list
 
+    return particle_list
+
+
+def keep_hadrons(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons in :code:`particle_list`.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.is_hadron() and not np.isnan(elem.is_hadron())
+        ]
+    return particle_list
+
+
+def keep_leptons(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only leptons in :code:`particle_list`.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.is_lepton() and not np.isnan(elem.is_lepton())
+        ]
+    return particle_list
+
+
+def keep_quarks(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only quarks in particle_list. This function can be used to filter all
+    gluons out if the input is a parton list.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.is_quark() and not np.isnan(elem.is_quark())
+        ]
+    return particle_list
+
+
+def keep_mesons(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only mesons in :code:`particle_list`.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.is_meson() and not np.isnan(elem.is_meson())
+        ]
+    return particle_list
+
+
+def keep_baryons(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only baryons in :code:`particle_list`.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.is_baryon() and not np.isnan(elem.is_baryon())
+        ]
+    return particle_list
+
+
+def keep_up(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing up quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_up() and not np.isnan(elem.has_up())
+        ]
+    return particle_list
+
+
+def keep_down(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing down quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_down() and not np.isnan(elem.has_down())
+        ]
+    return particle_list
+
+
+def keep_strange(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing strange quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_strange() and not np.isnan(elem.has_strange())
+        ]
+    return particle_list
+
+
+def keep_charm(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing charm quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_charm() and not np.isnan(elem.has_charm())
+        ]
+    return particle_list
+
+
+def keep_bottom(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing bottom quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_bottom() and not np.isnan(elem.has_bottom())
+        ]
+    return particle_list
+
+
+def keep_top(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Keep only hadrons containing top quarks in :code:`particle_list`.
+    This filter does not work for partons.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if elem.has_top() and not np.isnan(elem.has_top())
+        ]
+    return particle_list
+
+
+def remove_photons(particle_list: List[List[Particle]]) -> List[List[Particle]]:
+    """
+    Remove all photons from :code:`particle_list`.
+
+    Parameters
+    ----------
+    particle_list:
+        List with lists containing particle objects for the events
+
+    Returns
+    -------
+    list of lists
+        Filtered list of lists containing particle objects for each event
+    """
+    for i in range(0, len(particle_list)):
+        particle_list[i] = [
+            elem
+            for elem in particle_list[i]
+            if not np.isnan(elem.pdg) and int(elem.pdg) != 22
+        ]
     return particle_list
