@@ -8,19 +8,29 @@
 # ===================================================
 
 from abc import ABC, abstractmethod
+import numpy as np
+from typing import List, Union, Any, Tuple
+from sparkx.Particle import Particle
 
 
 class FlowInterface(ABC):
     @abstractmethod
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args: Any, **kwargs: Any):
         pass
 
     @abstractmethod
-    def integrated_flow(self, particle_data, *args, **kwargs):
+    def integrated_flow(
+        self, particle_data: List[List[Particle]], *args: Any, **kwargs: Any
+    ) -> Any:
         pass
 
     @abstractmethod
     def differential_flow(
-        self, particle_data, bins, flow_as_function_of, *args, **kwargs
-    ):
+        self,
+        particle_data: List[List[Particle]],
+        bins: Union[np.ndarray, List[float]],
+        flow_as_function_of: str,
+        *args: Any,
+        **kwargs: Any,
+    ) -> Any:
         pass
