@@ -9,6 +9,7 @@
 
 import numpy as np
 import random as rd
+from typing import Dict, List
 
 
 class GenerateFlow:
@@ -93,12 +94,15 @@ class GenerateFlow:
     `event generator <https://www.physik.uni-bielefeld.de/~borghini/Software/flow_analysis_codes/generator.cc>`__.
     """
 
-    def __init__(self, *vn, **vn_kwargs):
+    def __init__(self, *vn: float, **vn_kwargs: float) -> None:
+        self.n_: np.ndarray = np.array([])
+        self.vn_: np.ndarray = np.array([])
+
         if not vn and not vn_kwargs:
-            self.n_ = self.vn_ = None
+            self.n_ = self.vn_ = np.array([])
         else:
             try:
-                vn_dictionary = {
+                vn_dictionary: Dict[int, float] = {
                     int(kw.lstrip("v")): val for kw, val in vn_kwargs.items()
                 }
             except ValueError:
@@ -111,16 +115,20 @@ class GenerateFlow:
                 for k, v in enumerate(vn, start=2)
                 if v is not None and v != 0.0
             )
-            kwargs = dict(dtype=float, count=len(vn_dictionary))
-            self.n_ = np.fromiter(vn_dictionary.keys(), **kwargs)
-            self.vn_ = np.fromiter(vn_dictionary.values(), **kwargs)
 
-        self.phi_ = []
-        self.px_ = []
-        self.py_ = []
-        self.pz_ = []
+            self.n_ = np.fromiter(
+                vn_dictionary.keys(), dtype=int, count=len(vn_dictionary)
+            )
+            self.vn_ = np.fromiter(
+                vn_dictionary.values(), dtype=float, count=len(vn_dictionary)
+            )
 
-    def __distribution_function(self, phi):
+        self.phi_: List[float] = []
+        self.px_: List[float] = []
+        self.py_: List[float] = []
+        self.pz_: List[float] = []
+
+    def __distribution_function(self, phi: float) -> float:
         """
         Calculate the distribution function value for a given angle.
 
@@ -147,7 +155,7 @@ class GenerateFlow:
 
         return f * f_harmonic
 
-    def __sample_angles(self, multiplicity):
+    def __sample_angles(self, multiplicity: int) -> None:
         """
         Sample angles for a given multiplicity according to a given distribution.
 
@@ -161,7 +169,7 @@ class GenerateFlow:
             None
         """
         f_max = (1.0 + 2.0 * self.vn_.sum()) / (2.0 * np.pi)
-        phi = []
+        phi: List[float] = []
 
         while len(phi) < multiplicity:
             random_phi = rd.uniform(0.0, 2.0 * np.pi)
@@ -172,7 +180,7 @@ class GenerateFlow:
 
         self.phi_ = phi
 
-    def __thermal_distribution(self, temperature, mass):
+    def __thermal_distribution(self, temperature: float, mass: float) -> float:
         """
         Calculate the momentum magnitude from a thermal distribution.
 
@@ -188,7 +196,7 @@ class GenerateFlow:
             momentum_radial: float
                 The magnitude of the momentum.
         """
-        momentum_radial = 0
+        momentum_radial = 0.0
         energy = 0.0
         if temperature > 0.6 * mass:
             while True:
@@ -233,7 +241,9 @@ class GenerateFlow:
 
         return momentum_radial
 
-    def __sample_momenta_thermal(self, multiplicity, temperature, mass):
+    def __sample_momenta_thermal(
+        self, multiplicity: int, temperature: float, mass: float
+    ) -> None:
         """
         Sample momenta for a given multiplicity, temperature, and mass from
         a thermal distribution function.
@@ -276,7 +286,9 @@ class GenerateFlow:
         self.py_ = py
         self.pz_ = pz
 
-    def __artificial_pT_distribution(self, pT, pTmin, pT0, pT1, T0):
+    def __artificial_pT_distribution(
+        self, pT: float, pTmin: float, pT0: float, pT1: float, T0: float
+    ) -> float:
         """
         Calculate the artificial pT distribution.
 
@@ -319,7 +331,9 @@ class GenerateFlow:
             value = np.exp(-(pT1 - pT0) / T0) * (pT1 / pT) ** 7
         return value
 
-    def __artificial_flow_pT_shape(self, pT, pT0_bis, pT_sat, vn_sat):
+    def __artificial_flow_pT_shape(
+        self, pT: float, pT0_bis: float, pT_sat: float, vn_sat: float
+    ) -> float:
         """
         Calculate the artificial flow pT shape.
 
@@ -364,7 +378,9 @@ class GenerateFlow:
 
         return value
 
-    def __distribution_function_pT_differential(self, phi, vn_pT_list):
+    def __distribution_function_pT_differential(
+        self, phi: float, vn_pT_list: List[float]
+    ) -> float:
         """
         Calculates the pT-differential distribution function for a given
         azimuthal angle.
@@ -390,8 +406,11 @@ class GenerateFlow:
         return f_harmonic / f_norm
 
     def __create_k_particle_correlations(
-        self, multiplicity, k_particle_correlation, correlation_fraction
-    ):
+        self,
+        multiplicity: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate momentum components with k-particle correlations.
 
@@ -410,9 +429,9 @@ class GenerateFlow:
             Updates the internal arrays (px_, py_, pz_) with the
             generated correlated momenta.
         """
-        px = []
-        py = []
-        pz = []
+        px: List[float] = []
+        py: List[float] = []
+        pz: List[float] = []
         idx = 0
         while len(px) <= multiplicity:
             if rd.random() <= correlation_fraction:
@@ -432,8 +451,8 @@ class GenerateFlow:
         self.pz_ = pz
 
     def __generate_flow_realistic_pT_distribution(
-        self, multiplicity, reaction_plane_angle
-    ):
+        self, multiplicity: int, reaction_plane_angle: float
+    ) -> None:
         pTmax = 4.5
         pTmin = 0.1
         pT0 = 0.5
@@ -483,8 +502,8 @@ class GenerateFlow:
             self.pz_.append(rd.uniform(-1, 1) * pTmax)
 
     def generate_dummy_JETSCAPE_file(
-        self, output_path, number_events, multiplicity, seed
-    ):
+        self, output_path: str, number_events: int, multiplicity: int, seed: int
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta.
@@ -566,12 +585,12 @@ class GenerateFlow:
 
     def generate_dummy_JETSCAPE_file_realistic_pT_shape(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        random_reaction_plane=True,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution.
@@ -658,13 +677,13 @@ class GenerateFlow:
 
     def generate_dummy_JETSCAPE_file_multi_particle_correlations(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        k_particle_correlation,
-        correlation_fraction,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with random particle momenta resulting in
         the same flow for all transverse momenta. A fraction of k-particle
@@ -762,14 +781,14 @@ class GenerateFlow:
 
     def generate_dummy_JETSCAPE_file_realistic_pT_shape_multi_particle_correlations(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        k_particle_correlation,
-        correlation_fraction,
-        random_reaction_plane=True,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy JETSCAPE file with particles having flow with a more
         realistic transverse momentum distribution. A fraction of k-particle
@@ -871,8 +890,8 @@ class GenerateFlow:
             output.write("#	sigmaGen	0.0	sigmaErr	0.0")
 
     def generate_dummy_OSCAR_file(
-        self, output_path, number_events, multiplicity, seed
-    ):
+        self, output_path: str, number_events: int, multiplicity: int, seed: int
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with random particle momenta
         resulting in the same flow for all transverse momenta.
@@ -963,12 +982,12 @@ class GenerateFlow:
 
     def generate_dummy_OSCAR_file_realistic_pT_shape(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        random_reaction_plane=True,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution.
@@ -1064,13 +1083,13 @@ class GenerateFlow:
 
     def generate_dummy_OSCAR_file_multi_particle_correlations(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        k_particle_correlation,
-        correlation_fraction,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with random particle momenta
         resulting in the same flow for all transverse momenta. A fraction of
@@ -1177,14 +1196,14 @@ class GenerateFlow:
 
     def generate_dummy_OSCAR_file_realistic_pT_shape_multi_particle_correlations(
         self,
-        output_path,
-        number_events,
-        multiplicity,
-        seed,
-        k_particle_correlation,
-        correlation_fraction,
-        random_reaction_plane=True,
-    ):
+        output_path: str,
+        number_events: int,
+        multiplicity: int,
+        seed: int,
+        k_particle_correlation: int,
+        correlation_fraction: float,
+        random_reaction_plane: bool = True,
+    ) -> None:
         """
         Generate a dummy OSCAR2013 file with particles having flow with a more
         realistic transverse momentum distribution. A fraction of k-particle
