@@ -303,7 +303,7 @@ def test_filter_strangeness_in_Jetscape(tmp_path, particle_list_strange):
     tmp_jetscape_file = create_temporary_jetscape_file(tmp_path, 2)
     jetscape = Jetscape(tmp_jetscape_file)
     jetscape.particle_list_ = particle_list_strange
-    jetscape.strange_particles()
+    jetscape.keep_strange()
 
     assert np.array_equal(
         jetscape.num_output_per_event(), np.array([[1, 5], [2, 7]])
@@ -471,6 +471,15 @@ def test_Jetscape_print_with_empty_events(
     assert filecmp.cmp(jetscape_file_no_hadrons, output_path)
     os.remove(output_path)
 
+def test_Jetscape_print_with_no_events(
+        jetscape_file_path, output_path
+):
+    jetscape = Jetscape(jetscape_file_path)
+    jetscape.particle_list_ = [[], [], [], [], []]
+    jetscape.multiplicity_cut((100000000,None))
+    with pytest.warns(UserWarning):
+        jetscape.print_particle_lists_to_file(output_path)
+    os.remove(output_path)
 
 def test_Jetscape_get_sigmaGen(jetscape_file_path):
     jetscape = Jetscape(jetscape_file_path)
@@ -499,3 +508,29 @@ def test_Jetscape_read_parton_file(jetscape_file_path_partons):
     # Test construction with wrong particletype (not string)
     with pytest.raises(TypeError):
         Jetscape(jetscape_file_path_partons, particletype=1)
+
+def test_update_after_merge_warning(jetscape_file_path):
+    # Create two Jetscape instances with different particle types and defining strings
+    jetscape1 = Jetscape(jetscape_file_path)
+    jetscape1.particle_type_ = "parton"
+    jetscape1.particle_type_defining_string_ = "parton"
+    jetscape1.sigmaGen_ = (1,2)
+
+    jetscape2 = Jetscape(jetscape_file_path)
+    jetscape2.particle_type_ = "hadron"
+    jetscape2.particle_type_defining_string_ = "hadron"
+    jetscape2.sigmaGen_ = (2,2)
+
+    # Check if a UserWarning is issued when merging
+    with pytest.raises(TypeError):
+        jetscape1._update_after_merge(jetscape2)
+
+    with pytest.raises(TypeError):
+        jetscape1._update_after_merge(jetscape2)
+
+    jetscape2.particle_type_ = "parton"
+    jetscape2.particle_type_defining_string_ = "parton"
+
+    jetscape1._update_after_merge(jetscape2)
+
+    assert jetscape1.sigmaGen_ == (1.5,np.sqrt(2))
